requirements.txt
# Core dependencies
pymongo>=4.6.0
google-cloud-storage>=2.10.0
pandas>=2.1.0
pyarrow>=14.0.0
orjson>=3.9.0
redis>=5.0.0
tqdm>=4.66.0
psutil>=5.9.0
python-dotenv>=1.0.0

# Optional dependencies for enhanced performance
numpy>=1.24.0
fastparquet>=2023.10.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0
